app-history/navigate/goTo-detach-in-onnavigate.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="/common/blank.html"></iframe>
<script>
promise_test(async t => {
  await new Promise(resolve => window.onload = resolve);
  let iframe_constructor = i.contentWindow.DOMException;
  let i_win = i.contentWindow;
  i.contentWindow.history.pushState(1, "", "#1");
  assert_equals(i.contentWindow.appHistory.entries().length, 2);
  let key = i.contentWindow.appHistory.entries()[0].key;

  let onnavigateerror_called = false;
  i.contentWindow.appHistory.onnavigate = t.step_func(() => i.remove());
  i.contentWindow.appHistory.onnavigateerror = t.step_func(() => onnavigateerror_called = true);
  await promise_rejects_dom(t, 'AbortError', iframe_constructor, i.contentWindow.appHistory.goTo(key));
  assert_true(onnavigateerror_called);
}, "goTo() in a detached window");
</script>
